Clarify how to return ALL results for vision features
The vision.feature.Feature class constructor has parameter `max_results` which determines how many results to return. It defaults to 1 and can be set to other integers, but the docstring wasn't clear how to specify that all available results should be returned.

According to the implementation in the following method, a `max_results` value of `None` would effectively cause all available results to be returned. Therefore, this commit adds to the Feature class docstring to clarify this fact.

core.google.cloud.iterator.HTTPIterator#_has_next_page

```
    def _has_next_page(self):
        """Determines whether or not there are more pages with results.

        :rtype: bool
        :returns: Whether the iterator has more pages.
        """
        if self.page_number == 0:
            return True

        if self.max_results is not None:
            if self.num_results >= self.max_results:
                return False

        return self.next_page_token is not None
```